pyproject.toml
[project]
name = "valhalla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=24.10.0",
    "gitpython>=3.1.43",
    "rumps>=0.4.0",
]
